refactor(server): extract secure bounded-file input boundary
Move the owner-private bounded file read behind SecretSource::File into
a dedicated server::secure_file module that owns the platform seam the
Windows adapter will implement. The Unix policy moves verbatim; a
platform without a reviewed adapter now returns the typed
SecureFileError::Unavailable, which config maps to the same FileSecurity
error as before, so observable behavior is unchanged on every platform.
